==> 04_PaleoBio_GLMM_analysis/S01_EnvironmentalPCA_check.Rmd <==
---
title: "Supplementary Environmental PCA Check"
author: "FiG-T"
data: "`r Sys.Date()`"
output: 
  html_document: 
    keep_md: yes
    toc: yes
toc: TRUE
editor_options:
  markdown:
    wrap: 80
  chunk_output_type: inline
---

```{r libraries}
library(feather)
library(dplyr)
library(ggplot2)
library(stringr)
library(tidyr)
library(lme4)
```

```{r import data}
# import metafile with paleo variables 
gt_pc_ds_paleobio_meta <- feather::read_feather(
  path = "/data/feather/paleobio_meta_pc_ds_subset_full_08_2024.feather"
)

# mtDNA loci positioning 
mt_loci_pos <- feather::read_feather(
  path = "data/feather/mtDNA_loci_all_positions_classifications.feather"
)
```

```{r load palette}
source("https://raw.githubusercontent.com/FiG-T/climate_dist_mtDNA/refs/heads/main/climate_mtDNA_colour_palettes.R") # you may beed to add on a token here to gain access. 

```

```{r filter_data}
length(unique(gt_pc_ds_paleobio_meta$acc))
nrow(gt_pc_ds_paleobio_meta)

# remove rows without paleo data 
gt_pc_ds_paleobio_meta <- gt_pc_ds_paleobio_meta[,c(-31)] %>%
  filter(!is.na(mean_T_yr_paleo)) %>%
  distinct()
```

```{r run_envPCA}
# run a PCA on the environmental variables
env_pca <- prcomp(
  x = gt_pc_ds_paleobio_meta[,c(29,31:41)] %>% filter(!is.na(mean_T_yr_paleo)), 
  scale. = FALSE # values already scaled
)
# get summary
summary(env_pca)
# check length matches the number of rows   
length(env_pca$x[,1])

paleobio_meta_envpc <- gt_pc_ds_paleobio_meta 

# add in environmental PCs
paleobio_meta_envpc$envPC1 <- env_pca$x[,1]
paleobio_meta_envpc$envPC2 <- env_pca$x[,2]
paleobio_meta_envpc$envPC3 <- env_pca$x[,3]
paleobio_meta_envpc$envPC4 <- env_pca$x[,4]

# reorder columns 
paleobio_meta_envpc <- paleobio_meta_envpc %>%
  select(acc:lai_paleo,envPC1:envPC4, pos_154:pos_16592)

names(paleobio_meta_envpc)

write.csv(
  x = paleobio_meta_envpc, 
  file = "/data/genotypes_contempory_paleo_envpcs_raw.csv"
)
```


Using the `core_GLM_chi_scorer()` function as defined in the 02_run_paleobio_GLMMs.Rmd script. 

```{r run_GLMM}
# runnning glms to compare with 1st env PC
glm_scores_envpcs <- core_GLM_chi_scorer(
  input = paleobio_meta_envpc, 
  start_col = 46, 
  test_formula = "~ PC1 + PC2 + PC3 + PC4 + (1|iso2) + lat + precip_yr ", 
  null_formula = "~ PC1 + PC2 + PC3 + PC4 + (1|iso2) + lat + precip_yr + envPC1"
)

length(unique(glm_scores_envpcs$snp_pos))

glm_scores_envpcs_rcrs <- left_join(
  x = glm_scores_envpcs, 
  y = rCRS, 
  by = c("snp_pos"="aln_pos")
)

write.csv(
  x = glm_scores_envpcs_rcrs, 
  file = "/results/supp_glm_scores_envpcs_rCRS_03_2025.csv"
)
```

This can be plotted: 

```{r make_plot}

# making a figure
alpha <- 0.01

n_snps_envpc <-  length(unique(glm_scores_envpcs_rcrs$rCRS_pos)) # change depending on the output from format_core_GLM_scores
bon_threshold_envpc <- alpha/(n_snps*1)

ggplot2::ggplot(
  data = glm_scores_envpcs_rcrs, 
  mapping = aes(
    x = rCRS_pos,
    y = log10(`Pr(>Chisq)`)
  )
) +
  geom_point(
    colour = "lightpink3",
    size = 3
  ) +
  geom_line(
    y = -1*log10(bon_threshold_envpc), 
    colour = "black", 
    linewidth = 1,
  ) + 
  scale_x_continuous(
    breaks = seq(0,17000, 1000)
  ) +
 ylim(c(1, -10)
  ) +
  ylab(
    "Logged 10 P-value"
  ) +
  xlab(
    "Position (relative to rCRS)"
  ) +
  geom_segment(
    data = mt_loci_pos,
    ggplot2::aes(
      x = Starting, 
      #x = aln_start,
      xend = Ending, 
      #xend = aln_end,
      y = 0.5, 
      yend = 0.5, 
      col = classification
    ),
    linewidth = 3, 
    alpha = 0.7, 
    position = ggplot2::position_jitter(
      height = 0.1
    )
  ) +
  scale_colour_manual(
    values = c(
      mtDNA_palette
    )
  ) +
  #ggplot2::coord_polar(
  #) + 
  theme_minimal()+
  theme(
    legend.position = "none", 
    axis.line = ggplot2::element_line(
      linewidth = 1, 
      colour = "black"
    ), 
    plot.background = ggplot2::element_rect(
      fill = "white"
    ), 
    panel.grid = ggplot2::element_line(
      colour = "white", 
      linetype = "dotdash"
    ),
    panel.grid.major.x = ggplot2::element_line(
      colour = "snow4", 
      linetype = "dashed"
    ), 
    legend.background = ggplot2::element_rect(
      fill = "white"
    ), 
    axis.title = element_text(
      size = 20
    ),
    axis.text = element_text(
      size = 14
    )
  ) #+
#transparent_theme
```

```{r save_plot}
ggplot2::ggsave(
  filename = "/figures/supplementary/latprecip_vs_envpc1_manhattan_03_2025.png",
  plot = last_plot(),
  bg = "transparent", 
  width = 11, 
  height = 7
)
```
